# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: vatove REST API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
  /health/live:
    get:
      operationId: getLiveness
      responses:
        "200":
          description: Process is running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /health/ready:
    get:
      operationId: getReadiness
      responses:
        "200":
          description: PostgreSQL and Kafka are reachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
        "503":
          description: A dependency is unavailable
  /sync-runs:
    post:
      operationId: createSyncRun
      description: Queues the inclusive 60-day window ending on the current APP_TIMEZONE calendar date when no range is supplied.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                oldest: { type: string, format: date }
                newest: { type: string, format: date }
      responses:
        "202":
          description: Sync accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncRun"
        "400": { $ref: "#/components/responses/Problem" }
  /sync-runs/{id}:
    get:
      operationId: getSyncRun
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Sync state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncRun"
        "404": { $ref: "#/components/responses/Problem" }
  /activities:
    get:
      operationId: listActivities
      parameters:
        - $ref: "#/components/parameters/Oldest"
        - $ref: "#/components/parameters/Newest"
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
      responses:
        "200":
          description: Newest-first activity page
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/ActivitySummary" }
                  nextCursor: { type: [string, "null"] }
        "400": { $ref: "#/components/responses/Problem" }
  /activity-routes:
    get:
      operationId: listActivityRoutes
      parameters:
        - $ref: "#/components/parameters/Oldest"
        - $ref: "#/components/parameters/Newest"
      responses:
        "200":
          description: Simplified routes for activities in the requested calendar-date window
          content:
            application/geo+json:
              schema:
                $ref: "#/components/schemas/ActivityRouteCollection"
        "400": { $ref: "#/components/responses/Problem" }
  /activities/{id}:
    get:
      operationId: getActivity
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Normalized route and biometric samples
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityDetail"
        "404": { $ref: "#/components/responses/Problem" }
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Oldest:
      name: oldest
      in: query
      description: Inclusive oldest local calendar date. Must be supplied with newest; both default to the latest 60-day window when omitted.
      schema: { type: string, format: date }
    Newest:
      name: newest
      in: query
      description: Inclusive newest local calendar date. Must be supplied with oldest; both default to the latest 60-day window when omitted.
      schema: { type: string, format: date }
  responses:
    Problem:
      description: RFC 9457-style problem response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Health:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [ok, unavailable] }
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
    SyncRun:
      type: object
      required: [id, status, oldest, newest, discoveredCount, processedCount, failedCount, error, createdAt, startedAt, completedAt]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [queued, running, completed, failed] }
        oldest: { type: string, format: date }
        newest: { type: string, format: date }
        discoveredCount: { type: integer }
        processedCount: { type: integer }
        failedCount: { type: integer }
        error: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
        startedAt: { type: [string, "null"], format: date-time }
        completedAt: { type: [string, "null"], format: date-time }
    ActivitySummary:
      type: object
      required: [id, source, sourceActivityId, name, sport, startAt, movingTimeSeconds, distanceMeters, hasRoute, hasHeartRate, hasElevation]
      properties:
        id: { type: string, format: uuid }
        source: { type: string, const: intervals }
        sourceActivityId: { type: string }
        name: { type: string }
        sport: { type: string }
        startAt: { type: string, format: date-time }
        movingTimeSeconds: { type: [integer, "null"] }
        distanceMeters: { type: [number, "null"] }
        hasRoute: { type: boolean }
        hasHeartRate: { type: boolean }
        hasElevation: { type: boolean }
    ActivityDetail:
      allOf:
        - $ref: "#/components/schemas/ActivitySummary"
        - type: object
          required: [averageHeartRateBpm, maxHeartRateBpm, route, samples, heartRateZones]
          properties:
            averageHeartRateBpm: { type: [number, "null"] }
            maxHeartRateBpm: { type: [number, "null"] }
            route:
              oneOf:
                - $ref: "#/components/schemas/LineStringFeature"
                - type: "null"
            samples:
              type: array
              items: { $ref: "#/components/schemas/ActivitySample" }
            heartRateZones:
              type: array
              items: { $ref: "#/components/schemas/HeartRateZone" }
    LineStringFeature:
      type: object
      required: [type, geometry, properties]
      properties:
        type: { type: string, const: Feature }
        geometry:
          type: object
          required: [type, coordinates]
          properties:
            type: { type: string, const: LineString }
            coordinates:
              type: array
              minItems: 2
              items:
                type: array
                prefixItems: [{ type: number }, { type: number }]
                minItems: 2
                maxItems: 2
        properties:
          type: object
          required: [activityId]
          properties:
            activityId: { type: string, format: uuid }
    ActivityRouteCollection:
      type: object
      required: [type, features]
      properties:
        type: { type: string, const: FeatureCollection }
        features:
          type: array
          items: { $ref: "#/components/schemas/LineStringFeature" }
    ActivitySample:
      type: object
      required: [index, sourceIndex, longitude, latitude, elapsedSeconds, distanceMeters, elevationMeters, heartRateBpm, heartRateZone]
      properties:
        index: { type: integer, minimum: 0 }
        sourceIndex: { type: integer, minimum: 0 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        latitude: { type: number, minimum: -90, maximum: 90 }
        elapsedSeconds: { type: [number, "null"] }
        distanceMeters: { type: [number, "null"] }
        elevationMeters: { type: [number, "null"] }
        heartRateBpm: { type: [integer, "null"] }
        heartRateZone: { type: [integer, "null"] }
    HeartRateZone:
      type: object
      required: [index, label, color, minBpm, maxBpm, durationSeconds]
      properties:
        index: { type: integer, minimum: 0 }
        label: { type: string }
        color: { type: string }
        minBpm: { type: [integer, "null"] }
        maxBpm: { type: [integer, "null"] }
        durationSeconds: { type: [number, "null"], minimum: 0 }